+++ packages/amplify-dynamodb-simulator/__test__/index.test.js
@@ -90,4 +90,9 @@ describe('emulator operations', () => {
     emulators.push(emu);
     expect(emu.port).toBe(port);
   });
+
+  it('reports on invalid dbPath values', async () => {
+    expect.assertions(1);
+    await expect(ddbSimulator.launch({ dbPath: 'dynamodb-data' })).rejects.toThrow('invalid directory for database creation');
+  });
 });