+++ packages/graphql-transformer-common/src/PredictionsResourceIDs.ts
@@ -1,11 +1,11 @@
 export class PredictionsResourceIDs {
   static actionMapID = 'predictionsActionMap';
   static iamRole = 'predictionsIAMRole';
-  static lambdaIAMRole = 'predictionsLambdaIAMRole'
+  static lambdaIAMRole = 'predictionsLambdaIAMRole';
   static lambdaName = 'predictionsLambda';
-  static lambdaID = 'predictionsLambdaFunction'
+  static lambdaID = 'predictionsLambdaFunction';
   static lambdaHandlerName = 'predictionsLambda.handler';
-  static lambdaRuntime = 'nodejs10.x';
+  static lambdaRuntime = 'nodejs14.x';
   static lambdaTimeout = 60;
 
   static getPredictionFunctionName(action: string) {