+++ packages/graphql-transformers-e2e-tests/src/LambdaHelper.ts
@@ -19,7 +19,7 @@ export class LambdaHelper {
         Code: {
           ZipFile: zipContents,
         },
-        Runtime: 'nodejs10.x',
+        Runtime: 'nodejs14.x',
         Handler: `${filePrefix}.handler`,
         Role: roleArn,
       })