+++ packages/amplify-category-hosting/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@2.6.2...amplify-category-hosting@2.6.3) (2021-02-26)
+
+**Note:** Version bump only for package amplify-category-hosting
+
+
+
+
+
 ## [2.6.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@2.6.1...amplify-category-hosting@2.6.2) (2021-02-24)
 
 **Note:** Version bump only for package amplify-category-hosting