+++ packages/amplify-util-mock/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.28.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.28.1...amplify-util-mock@3.28.2) (2021-02-26)
+
+
+
+## 4.44.2 (2021-02-26)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
 ## [3.28.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@3.28.0...amplify-util-mock@3.28.1) (2021-02-24)